fix: 챌린지의 타입이 URL에 노출되는 현상 수정
