+++ src/routes.tsx
@@ -3,7 +3,7 @@ import DefaultLayout from './components/common/layouts/DefaultLayout';
 import FamilyPageLayout from './components/common/layouts/FamilyPageLayout';
 import PrivateLayout from './components/common/layouts/PrivateLayout';
 import routes from './constants/routes';
-import ChallengDetailPage from './pages/ChallengDetailPage';
+import ChallengeDetailPage from './pages/ChallengeDetailPage';
 import ChallengeListPage from './pages/ChallengeListPage';
 import FamilyCreatePage from './pages/FamilyCreatePage';
 import FamilyEntryPage from './pages/FamilyEntryPage';
@@ -43,7 +43,7 @@ const privateChildren: RouteObject[] = [
   {
     // 챌린지 상세/참여 페이지
     path: `${routes.challenge}/:challengeId`,
-    element: <ChallengDetailPage />,
+    element: <ChallengeDetailPage />,
   },
   {
     // 마이 페이지
